refactor: unify agent registration
